squash commits

+++ electron-app/src/utils.ts
@@ -20,7 +20,15 @@ import {
   MAINNET_BASE_FOLDER,
   REGTEST_BASE_FOLDER,
   TESTNET_BASE_FOLDER,
+  BLK_FILE,
+  REV_FILE,
+  TESTNET_BASE_FOLDER_REINDEX,
+  MAINNET_BASE_FOLDER_REINDEX,
 } from './constants';
+import { DAT_FILE_TYPE } from '@defi_types/fileExtensions';
+import * as log from '././services/electronLogger';
+import { IPCResponseModel } from '@defi_types/common';
+import { ADDNODE, MASTERNODE_OPERATOR } from '../../typings/rpcConfig';
 
 export const getPlatform = () => {
   switch (platform()) {
@@ -38,26 +46,17 @@ export const getPlatform = () => {
   }
 };
 
-export const getBinaryParameter = (obj: any = {}) => {
-  let remote: any = {
-    rpcallowip: '',
-    rpcauth: '',
-    rpcport: 0,
-    rpcuser: '',
-    rpcpassword: '',
-    rpcbind: '',
-    datadir: '',
-  };
-  remote.rpcallowip = '0.0.0.0/0';
-  if (!!obj && Array.isArray(obj.remotes)) {
-    remote = Object.assign({}, remote, obj.remotes[0]);
-    remote.rpcbind = obj.remotes[0].rpcconnect;
-    delete remote.rpcconnect;
-  }
-  return Object.keys(remote).map((key) => `-${key}=${remote[key]}`);
+export const createResponseMessage = <T>(
+  success: boolean,
+  res: T
+): IPCResponseModel<T> => {
+  return { success, data: res };
 };
 
-export const responseMessage = (success: boolean, res: any) => {
+export const responseMessage = <T>(
+  success: boolean,
+  res: T
+): IPCResponseModel<T> => {
   if (success) {
     return { success: true, data: res };
   }
@@ -81,7 +80,22 @@ export const createDir = (dirPath: string) => {
 
 // Get file data
 export const getFileData = (filePath: string, format: string = 'utf-8') => {
-  return fs.readFileSync(filePath, format);
+  const fileData = fs.readFileSync(filePath, format);
+  return formatConfigFileRead(fileData);
+};
+
+// Add squarebrackets masternode_operator in config file
+export const formatConfigFileRead = (fileData: string) => {
+  return fileData
+    .replace(new RegExp(MASTERNODE_OPERATOR, 'gi'), `${MASTERNODE_OPERATOR}[]`)
+    .replace(new RegExp(ADDNODE, 'gi'), `${ADDNODE}[]`);
+};
+
+// Remove squarebrackets masternode_operator in config file
+export const formatConfigFileWrite = (fileData: string) => {
+  return fileData
+    .replace(/masternode_operator[\[\]']+/g, MASTERNODE_OPERATOR)
+    .replace(/addnode[\[\]']+/g, ADDNODE);
 };
 
 // write / append on UI config file
@@ -113,7 +127,7 @@ export const getRpcAuth = (rpcuser: string) => {
   };
 };
 
-export const getProcesses = (args: any) => {
+export const getProcesses = (args: any): Promise<any> => {
   return new Promise((resolve, reject) => {
     ps.lookup(args, (err: any, result: unknown) => {
       if (err) return reject(err);
@@ -123,12 +137,22 @@ export const getProcesses = (args: any) => {
 };
 
 export const stopProcesses = (processId: number | string) => {
-  return new Promise((resolve, reject) => {
-    ps.kill(processId, 'SIGTERM', (err: any, result: unknown) => {
-      if (err) return reject(err);
-      return resolve(result);
-    });
-  });
+  return getProcesses({ pid: parseInt(processId?.toString(), 10) }).then(
+    (processes: any[]) => {
+      if (processes != null && processes[0] != null) {
+        return new Promise((resolve, reject) => {
+          ps.kill(processId, 'SIGTERM', (err: any, result: unknown) => {
+            if (err) return reject(err);
+            return resolve(result);
+          });
+        });
+      } else {
+        return new Promise((resolve, reject) => {
+          return resolve(true);
+        });
+      }
+    }
+  );
 };
 
 export function sleep(ms: number) {
@@ -183,4 +207,67 @@ export const getBaseFolder = () => {
     baseFolder = REGTEST_BASE_FOLDER;
   }
   return baseFolder;
+};
+
+export const getBaseFolderReindex = () => {
+  const data = getIniData(CONFIG_FILE_NAME);
+  let baseFolder = MAINNET_BASE_FOLDER_REINDEX;
+  if (data.testnet && parseInt(data.testnet, 10)) {
+    baseFolder = TESTNET_BASE_FOLDER_REINDEX;
+  }
+  if (data.regtest && parseInt(data.regtest, 10)) {
+    baseFolder = REGTEST_BASE_FOLDER;
+  }
+  return baseFolder;
+};
+
+export const deletePeersFile = () => {
+  try {
+    const baseFolder = getBaseFolderReindex();
+    const destFileName = `peers.dat`;
+    const destFilePath = path.join(baseFolder, destFileName);
+    if (checkPathExists(destFilePath)) {
+      deleteFile(destFilePath);
+      log.info(`Deleted peers file in ${destFilePath}`);
+    }
+  } catch (error) {
+    log.error(error);
+  }
+};
+
+export const deleteBlocksAndRevFiles = () => {
+  try {
+    log.info('Starting Delete Block and Rev Files...');
+    const baseFolder = getBaseFolderReindex();
+    const destFolder = path.join(baseFolder, 'blocks');
+    fs.readdirSync(destFolder).forEach((file) => {
+      const blkFile = path.join(destFolder, file);
+      if (
+        file?.endsWith(DAT_FILE_TYPE) &&
+        (file?.includes(BLK_FILE) || file?.includes(REV_FILE))
+      ) {
+        log.info(`Deleting ${blkFile}...`);
+        if (checkPathExists(blkFile)) {
+          deleteFile(blkFile);
+        }
+      }
+    });
+    log.info('Delete Block and Rev Files completed...');
+  } catch (error) {
+    log.error(error);
+  }
+};
+
+export const deleteBanlist = () => {
+  try {
+    const baseFolder = getBaseFolderReindex();
+    const destFileName = `banlist.dat`;
+    const destFilePath = path.join(baseFolder, destFileName);
+    if (checkPathExists(destFilePath)) {
+      deleteFile(destFilePath);
+      log.info(`Deleted banlist file in ${destFilePath}`);
+    }
+  } catch (error) {
+    log.error(error);
+  }
 };
